+++ backend/package.json
@@ -6,10 +6,12 @@
     "start": "node dist/index.js",
     "build": "tsc",
     "test": "jest",
+    "dev": "nodemon src/index.ts",
+    "test:coverage": "jest --coverage",
     "test:unit": "jest tests/unit",
     "test:integration": "jest tests/integration",
-    "test:coverage": "jest --coverage",
-    "dev": "nodemon src/index.ts"
+    "test:unit:tasks": "jest tests/unit/tasks",
+    "test:integration:tasks": "jest tests/integration:tasks"
   },
   "keywords": [],
   "author": "",
