+++ backend/src/validators/tasks/deleteTaskValidator.ts
@@ -1,0 +1,5 @@
+import { param } from "express-validator";
+
+export const deleteTaskValidator = [
+  param("id").isInt().withMessage("Task ID must be an integer").toInt(),
+];
