+++ backend/src/validators/tasks/markTaskAsDoneValidator.ts
@@ -1,0 +1,7 @@
+import { param } from "express-validator";
+
+export const markTaskAsDoneValidator = [
+    param("id")
+        .isInt().withMessage("Task ID must be an integer")
+        .toInt(),
+];
